requirement.txt
torch>=1.13.0
openai-whisper>=20231106
webrtcvad>=2.0.10
pyaudio>=0.2.11
pvporcupine>=1.9.5
pvrecorder>=1.1.1
piper-tts>=1.2.0
gradio>=4.0.0
requests>=2.28.0
numpy>=1.21.0
librosa>=0.9.0
psutil>=5.9.0
